Move duke exception file into exception sub package in duke
