+++ src/main/java/duke/TaskList.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.exceptions.DukeException;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
